// openapi/models/create-matter-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateMatterRequest
 */
export interface CreateMatterRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateMatterRequest
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof CreateMatterRequest
     */
    'description': string;
    /**
     * 
     * @type {number}
     * @memberof CreateMatterRequest
     */
    'tenantid'?: number;
}